+++ src/selectize.js
@@ -1082,6 +1082,12 @@ $.extend(Selectize.prototype, {
 
 			for (j = 0, k = optgroups && optgroups.length; j < k; j++) {
 				optgroup = optgroups[j];
+				if (!self.optgroups.hasOwnProperty(optgroup) && typeof self.settings.optgroupRegister === 'function') {
+					var regGroup;
+					if (regGroup = self.settings.optgroupRegister.apply(self, [optgroup])) {
+						self.registerOptionGroup(regGroup);
+					}
+				}
 				if (!self.optgroups.hasOwnProperty(optgroup)) {
 					optgroup = '';
 				}