+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 <!-- Feel free to put either your handle and/or full name, according to
      your privacy needs -->
+* Fix issues with required attribute in MS Edge (#1320)
+
+    *@TomOne*
 
 *  New feature: allow to disable single options or complete optgroups